Housekeeping - Remove clock duplication
We are going to add a new helper as part of this change that returns today's date without the time element. There are a number of places we do this already so its about time we brought them into one place.

This means we'll have another test for GeneralLib where we want to control the time. So, before we add it we figured it was time to remove some of the duplication and handle the clock variables and cleanup in one place.