Remove duplicate tests in `ReturnLogsSetupController`
We fixed a failing unit test following our changes to `units.njk`, but we didn't notice at the time that there was a second failing unit test in the same file. On further inspection, several endpoint tests are duplicated, we assume due to merge issues at some point. We therefore remove them.