Add new today() helper to GeneralLib
https://eaflood.atlassian.net/browse/WATER-5144

We got started on a change to the service that would [Allow ended 'Not due yet' returns to be submitted](https://github.com/DEFRA/water-abstraction-system/pull/2418).

As is often the case, we needed to compare a date from the return log to the current date, but the current date without the time.

This is something we frequently have to do to ensure we are comparing 'date-to-date' rather than 'date-to-datetime'.

If not we hit edge-case errors when when the date we're comparing matches the current date.

We were going to do it as some housekeeping within the original ticket, but then we clocked how often we've needed to do this, so the change became something in its own right.