Save return log ids for the return forms journey (#2421)
https://eaflood.atlassian.net/browse/WATER-5226

We recently added a migration to add a return log ids column to notification view.

The returns log ids are currently stored in the metadata column.

We want to make these easily accessible so we are adding a 'return_log_ids' column.

This will store the return log ids for a notification as a json array e.g ["a82c1a23-d7ac-4410-8534-acccab16b850"].

Some notifications are linked to only one return log id (such as the paper forms) but others are linked to multiple return log ids. This is why we use the jsonb field and saved the ids in an array.

We only want to store the return ids for the return forms at the moment.

We will be making an effort in the future to save the return ids for invitations and reminders.