Allow ended 'Not due yet' returns to be submitted (#2418)
https://eaflood.atlassian.net/browse/WATER-5144

The business has needed to revoke a licence, which results in a split return log. Basically, its return period was the whole cycle (2025-04-01 to 2026-03-31), which means its due date was set 28 days after both the period and the cycle end (2026-04-28).

When the licence gets revoked, the return log is replaced with one whose end date matches the licence end date, for example, 2025-09-19.

However, the return log will still be linked to the same cycle, so that the due date will remain calculated as 2026-04-28.

In the service, it will display as **NOT DUE YET**. When we added this status and took over returns management, the business decision was that it should not be possible to submit these return logs.

But now they have realised they need to. They want to capture the return submission data now and not have to wait till April 2026!

This change amends the logic for the action button. Rather than only showing the **Submit** button when we hit 28 days before the due date, we'll make it available as soon as the current date is greater than the return log's end date.