Fix date filter in view notices excluding records
https://eaflood.atlassian.net/browse/WATER-5271

The view notices page has a Sent dates filter that allows users to filter notices by

- Date from
- Date to

An issue has been found when using the 'Date to' filter: any notices that match the date entered are excluded from the results.

The problem is we are filtering by a date, against a datetime field.

Specifically, `created_at`` stores both the date and time the record was created, for example, '2025-09-12T09:13:26.924Z'.

The date we are using to filter by does not have a time element. So, it is essentially '2025-09-12T00:00:00.000Z'.

The `WHERE` clause is `WHERE created_at <= toDate`, which means those created on the filter date are being excluded.

It's a relatively simple fix. We need to alter `toDate`` to be '2025-09-12T23:59:59.999Z' i.e. the very last point in the day.