Update bas format validation result presenter to handle multiple errors with the same field (#2427)
A single field could have multiple errors. When this is the case we want to set the first error only. This assumes the validation order is correct. For example, A licence number must be valid in order to check it has any due returns. Therefore, the validator should order them appropriately.

This change updates the bse presenter 'formatValidationResult' to handle multiple errors for the same field.

This was spotted in the ad hoc journey when selecting a licence. We were seeing two errors in the error summary and only one for the field.