Stop nil returns from creating lines (#2429)
https://eaflood.atlassian.net/browse/WATER-5278

When a nil return is submitted the `return_submission_lines` table is incorrectly being populated.

This is happening because the function that creates the lines checks the session to see if there is any line data in it, if there is no line data no lines are created. The problem is that there is always line data in the session because:

- When a session is initiated for entering a return for the first time empty lines are added to the session.
- When editing a return, if that return isn't a nil return, lines are added to the session and remain there even if the return is updated to be a nil return.

In this PR we will ensure that no lines are created when the submission is a nil return by updating the logic in the `CreateReturnLinesService` to check for a nil return rather than check if there are any return lines.