Add a status check for emails after sending a batch notification (#2430)
* Add a status check for emails after sending a batch notification

Ticket -

We received feedback from users / the business about needing to see abstraction alerts status updates as soon as possible.

Previously we ran a status update job at the end of the day to update all these statuses.

We need to update this for the use case above. We can extend our existing status update  services to handle this for us.

This change updates the Notify update status services to receive an event id, this event id (when present) will be used as a trigger to only fetch notifications for that notice event (and only for emails).

An issue with this approach is the possibility of reaching the rate limit, as we are now potentially calling notify twice (with the same batch count).

This change also updates the batch by reducing it in half, from 250 to 125. This is to allow for a situation where we send 125 emails and need check the status for all 125 emails.

The overall batch limit is the same at 250. There are of course cases where we only send letters, and we only batch 125 a time. For the sake of simplicity this feels appropriate.