Revert abstraction alerts to send to licence holder as well as the additional contacts (#2432)
https://eaflood.atlassian.net/browse/WATER-5283

We initially implemented alerts to send a notice to the licence holder and additional contacts, this resembled the legacy codes implementation. We were then told this is wrong... so we updated the logic to only send to an additional contact if present, this was done here - #2077.

We have told again that the first implementation was correct and the change made to only send to additional contacts should be changed back.

This change does that. We can not straight up revert the commit and some things have been changed in other places.

So this change re-implements the original logic.