Housekeeping - Refactor fetch service to new way
This service is one of the first we added, when creating numerous DB records was not an issue because we cleaned the DB between each test.

With us now only cleaning at the start of the test run, in preparation for moving to a testing framework that will run concurrently, we now focus on limiting our DB interactions.

For example, this refactoring means we go from creating 8 bill runs to just 2. We also clean up after ourselves, to reduce the chance of what we leave impacting other unit tests.