Seed missing digitise users
https://github.com/DEFRA/water-abstraction-team/issues/115

In the service when creating an internal user there are eight sets of permissions that can be assigned.

- Basic access (View licences and returns)
- Billing and Data (Send returns communications, monitor statistics and edit data)
- Environment Officer (Send hands-off flow notifications)
- National Permitting Service (Send renewals)
- National Permitting Service and Digitise! editor (Send renewals and digitise licence information)
- National Permitting Service and Digitise! approver (Send renewals, digitise licence information and approve changes)
- Permitting and Support Centre (Send renewals)
- Waste and Industry Regulatory Service (Process returns)

As part of making working with the service easier for the team, we added supporting for seeding a database with records.

It's mainly used for our test suite, but it also comes in handy for seeding environments with various test users.

Only we've spotted that we don't have users for two of the permission sets: Digitise! editor and Digitise! approver.

This changes adds those to our user seeding.