Remove the testing of the feature flag
My guidance for the team is when writing the unit tests, if a feature flag is involved assume the feature is enabled. The flag is intended to be temporary, whereas the feature should be around for the foreseeable future.

Therefore write your tests as if this is 'the norm'. Your own and the QA's exploratory testing is sufficient for ensuring something stays hidden if we are not quite ready for it to go live.