Alternate presenter test strategy
When we talked about using iteration and the unit tests I had envisaged a something simpler, for example, given these scopes we get these results.

What I realise from looking at the unit tests is it could never be that simple. You have to somehow define the relationship between the scopes w're passing in and the results expected.

I'm guessing that is what `SECTIONS_TO_TEST` and `LINKS_TO_TEST` is trying to solve, as well as supporting the drive to support the iteration of the two 'tests' being executed.

The fact I'm guessing, and that my little brain needed to grab the pen and paper to muddle my way through the various `Object.entries()` and `new Set()` calls are the reasons I wanted to park this and have a think.

My alternate strategy approaches the problem from what someone can tangibly see in the service; the types of users (or permission sets) you can see when adding or editing a user.

Both me as a dev, and the QA checking the work, are more likely to understand permissions at the user level, rather than the scope level we have to deal with in the presenter. It's also the way both of us will manually test the page.

Yes, there is definitely repetition. But the end result is only a few more lines of test code. And in my opinion the tests are easier to parse.

> "Given I'm an X type user, the presenter will allow Y sections and Z links to be displayed".

More than happy to continue the conversation on this. Else if you are happy, we can merge the change and then get the new page into QA.