Update service test to assert whole object
Typically we'll assert properties when we expect their behaviour to change under different scenarios.

In this case its all covered by the presenter tests, as is the case for 99% of these kinds of view-foo.service modules.

So, normally we'd just add a basic test that asserts the whole object we get back from the service.