Fix broken links in the view
If you don't include the leading slash, the browser automatically works it out.

For example the 'Send flow notices' restriction link was this; `notifications/1?start=1`

When you go to click it, the browser has transformed it into `http://localhost:8008/system/notifications/1?start=1`. Adding the leading slash tells it to do this `http://localhost:8008/notifications/1?start=1`, which means it now works!