Prevent selecting default abstraction period outside of return period (#2431)
https://eaflood.atlassian.net/browse/WATER-5277

When a user selects a period for a return, they are presented with a default option, which is the defined return period from the licence for that return, for example 1 November to 31 March.

However, sometimes this falls outside of the period of the return, which may happen when the licence is revoked.

In this case, we should not display this default option - at the moment this then causes an error when the option is submitted.

This change checks to see if there are any abstraction periods that overlap with the return period and if there aren't then it doesn't display the default return period option.