Suggested contributions to the manage page changes (#2436)
Suggested contributions to the manage page changes

Sometimes it is easier to show rather then tell. These are my suggested changes to the PR for consideration.

* Alternate presenter test strategy

When we talked about using iteration in the unit tests, I had envisioned something simpler; given these scopes, we get these results.

What I realise from looking at the unit tests is that it could never be that simple. You have to somehow define the relationship between the scopes we're passing in and the results expected.

I'm guessing that is what `SECTIONS_TO_TEST` and `LINKS_TO_TEST` are trying to solve, as well as supporting the drive to support the iteration of the two 'tests' being executed.

The fact I'm guessing, and that my little brain needed to grab the pen and paper to muddle my way through the various `Object.entries()` and `new Set()` calls, are the reasons I wanted to park this and have a think.

My alternative strategy approaches the problem from what someone can tangibly see in the service: the types of users (or permission sets) that are visible when adding or editing a user.

Both I, as a developer, and the QA team checking the work are more likely to understand permissions at the user level, rather than the scope level we have to deal with in the presenter. It's also how we will manually test the page.

Yes, there is definitely repetition. But the result is only a few more lines of test code. And in my opinion, the tests are easier to parse.

> "Given I'm an X type user, the presenter will allow Y sections and Z links to be displayed".

More than happy to continue the conversation on this. Else if you are happy, we can merge the change and then get the new page into QA.

* Remove the testing of the feature flag

My guidance for the team is that when writing the unit tests, if a feature flag is involved, assume the feature is enabled. The flag is intended to be temporary, whereas the feature should remain in place for the foreseeable future.

Therefore, write your tests as if this is 'the norm'. Your own and the QA's exploratory testing is sufficient for ensuring something stays hidden if we are not quite ready for it to go live.

* Update service test to assert whole object

Typically, we'll assert properties when we expect their behaviour to change under different scenarios.

In this case its all covered by the presenter tests, as is the case for 99% of these kinds of view-foo.service modules.

So, we'd usually add a basic test that asserts the entire object we receive from the service.

* Fix broken links in the view

If you don't include the leading slash, the browser automatically works it out.

For example the 'Send flow notices' restriction link was this; `notifications/1?start=1`

When you go to click it, the browser has transformed it into `http://localhost:8008/system/notifications/1?start=1`. Adding the leading slash tells it to do this `http://localhost:8008/notifications/1?start=1`, which means it now works!

* Housekeeping

- Update .env.example to match current
- Order feature flags alphabetically
- Update view to "template standard"