Return logs - Persist zero quantities (#2433)
https://eaflood.atlassian.net/browse/WATER-5279

When submitting a return. If the line quantity is entered as zero it is not being saved to the DB when the return is submitted. Instead it is being recorded as `null`. Also zeros are not being persisted to the session when editing a return.

This was happening because if the quantity was `null`, `undefined` or zero, undefined was being returned. This has been fixed in this PR by just returning the quantity instead of `undefined` if the quantity is not truthy.