Prevent empty return submissions https://eaflood.atlassian.net/browse/WATER-5192
Currently it is possible to step through the return submission process 'enter & submit' but to not enter any volumes/meter readings and then press confirm.  The total return volume appears as 0m3 but all the return lines within the submitted return are blank.

This is not desirable behaviour as these type of returns should be recorded as nil returns.

In this PR validation is going to be added to the check page to prevent the submission of returns with a zero volume.