Update `multiple-entries.njk` to use new template
Note that this required a small change to `MultipleEntriesValidator`. The existing validation was wrapped in an object to ensure we return a consistent anchor id for the error summary links (otherwise, the `path` value we use to determine the anchor would be things like the error value's position in the validated array and not the `multipleEntries` id of the component).