Show under query btn for ended 'Not due yet' rtns
https://eaflood.atlassian.net/browse/WATER-5287

The business has needed to revoke a licence, which results in a split return log. Basically, its return period was the whole cycle (2025-04-01 to 2026-03-31), which means its due date was set 28 days after both the period and the cycle end (2026-04-28).

When the licence gets revoked, the return log is replaced with one whose end date matches the licence end date, for example, 2025-09-19.

However, the return log will still be linked to the same cycle, so that the due date will remain calculated as 2026-04-28.

In the service, it will display as **NOT DUE YET**. When we added this status and took over returns management, the business decision was that it should not be possible to submit these return logs.

But now they have realised they need to. They want to capture the return submission data now and not have to wait till April 2026!

We updated the presenter logic in [Allow ended 'Not due yet' returns to be submitted](https://github.com/DEFRA/water-abstraction-system/pull/2418) to make the button visible. But because it wasn't visible on the page, we overlooked the **Mark as under query** button.

If you can submit or edit a return log, you should also be able to mark it as under query.