Refactor 'end' process in setup notice journey - create notifications before batching (#2442)
* Refactor 'end' process in setup notice journey - create notifications before batching

https://eaflood.atlassian.net/browse/WATER-5280

Currently, when a user clicks Confirm on the check-recipients page, as part of sending a notice, we follow this process.

Convert the notice setup and recipient data into package to be sent to Notify

We send the package to Notify

We insert the result and the data into the DB as a ‘notification’.

Because there may be many notifications to send, or in the case of return forms, each can take some time to process, we do this behind the scenes to avoid users seeing any spinning wheels.

So, they will be immediately taken to the ‘Confirmation’ page, which includes a link to view the notice.

If the user is keen, and the process is going to take a moment to complete, when they view the notice and its notifications, the page will be empty. If they were to wait a moment and refresh, then the records would appear. However, we do not consider this a reasonable workaround.

This change is the initial refactored to fix this problem.

We have lifted out creating the notifications into the submit check service (in preparation to save them before batching, which we will do in another change).

This affects the tests leading to a complete rewrite. We can remove any concept of a journey / session and focus purely on sending and saving notifications.

We have also updated the event model to process the 'licences' column as JSON. This simplifies both logic and tests.