Refactor the reference code to only be used by the batch notification service
https://eaflood.atlassian.net/browse/WATER-5280

We are refactoring the batch notification.

As part of this refactor, we have identified an improvement to the way we compose notifications.

Previously every notification was setting the reference itself (also known as 'referenceCode').

This is uniquely generated to link all the records in Notify. We do not use this locally other for than display. We link notations to a notice through the notice id (even id).

So we can remove the need for notifications to 'know' about it.

This change removes setting the reference code in the various notification presenters. The batch service now has the responsibility of setting the reference code.

This was found as we were previously deleting the reference before saving a new notification. As we are currently in the process of making changes to save the notification data before sending to notify.