Fixing tests - by deleting them!
The presenter is just calling `formatReturnLogStatus()` in the base presenter. It is not manipulating the return log or performing any other logic around it.

In these cases we know we have the functionality covered because of the tests we have for `formatReturnLogStatus()`. The tests we have here (once we updated them) would just be duplicates.

Therefore it is better to delete them, rather than maintaining two sets of the same tests.