Migrate Manage page (#2283)
https://eaflood.atlassian.net/browse/WATER-5206

Migrate the `/manage` page from `water-abstraction-ui` to `water-abstraction-system`.

Add a new new feature flag `ENABLE_SYSTEM_MANAGE_VIEW` to control whether this is displayed here. This flag will also need to be added to `water-abstraction-ui` and assigned the same value.

The old `/manage` page had a number of links to other areas of the system that were controlled by feature flags. Most of these feature flags that are in `water-abstraction-ui` are now redundant, so the only one that needs to be copied/included in `water-abstraction-system` is `ENABLE_AD_HOC_NOTIFICATIONS`