Add an update notifications service
https://eaflood.atlassian.net/browse/WATER-5280

We currently create notifications after sending them to Notify. This allowed us to save the notification and the Notify response in one query.

We are changing this approach after seeing some delays in showing a notification on the notice page.

We want to create the notification and mark them as 'pending', and then update the notification with the Notify response.

This change adds the logic update a notification with the Notify response.

The responses can be either successful or unsuccessful. This is handled in the 'NotifyUpdatePresenter'.