Determine relevant lic vers where start date same
https://eaflood.atlassian.net/browse/WATER-5251

The issue that was raised was a licence that had two licence versions with the same start date. The first covered a single day: 1 April 2025 and originally had the wrong purposes linked to it.

Rather than delete it, the users created a second licence version that also started on 1 April 2025 (hence the previous only covering a single day).

When setting up the return version, the user selected 1 April 2025 as the start date to use. This resulted in `FetchRelevantLicenceVersionService` returning the earlier (single day) licence version rather than the later one. This is because the results are ordered by `end date DESC` and prior to this change, the service just return the first result.

> If they had selected 2 April 2025 we wouldn't have an issue! 😩

The follow examples explain the problem in more detail.

**Both superseded**

|Issue|Increment|Start date|End data  |Status    |
|-----|---------|----------|----------|----------|
|100  |0        |2025-04-01|2025-04-01|superseded|
|101  |0        |2025-04-01|2025-06-30|superseded|

In this example both licence versions are superseded so both have an end date. If the selected start date is _after_ 2024-04-01 then issue 100 would not be in the results and we'll take issue 101 as the first record.

However, if the selected start date is 2025-04-01, both would be returned because the filter is 'end date >= selected start date'.

Our results are sorted by end date ascending, which means issue 100 will be the first result. But as it is replaced by issue 101, that's the 'relevant' licence version we want to return. In this case we return the last record of those in with the same date.

**One current, one superseded**

|Issue|Increment|Start date|End data  |Status    |
|-----|---------|----------|----------|----------|
|100  |0        |2025-04-01|2025-04-01|superseded|
|101  |0        |2025-04-01|          |current   |

In this example, we have a current licence version and a superseded one. Again, if the selected start date is _after_ 2024-04-01 then issue 100 would not be in the results and we'll take issue 101 as the first record.

However, if the selected start date is 2025-04-01, both would be returned.

When you sort ascending, null values come last. So again, the record we want to return is now the last one, not the first.

---

This change updates the logic in the fetch to return all records (doing this all in SQL resulted in a gnarly script!) and then perform extra processing if we are dealing with a 'same start date' problem.