Update create notifications test
https://eaflood.atlassian.net/browse/WATER-5280

We currently create notifications after sending them to Notify. This allowed us to save the notification and the Notify response in one query.

We are changing this approach after seeing some delays in showing a notification on the notice page.

We want to create the notification and mark them as 'pending', and then update the notification with the Notify response.

The update logic will be implemented in another change.

This change updates the create notifications tests to only test multiple inserts (this was raised when working on https://github.com/DEFRA/water-abstraction-system/pull/2451).