Support for stabilising the acceptance tests
https://eaflood.atlassian.net/browse/WATER-5294

With all the changes to notices, beginning work on dynamic due dates, and our efforts to clean up and standardise the UI pages (behind the scenes known as “template standard”), the tech team is no longer able to run the acceptance tests reliably.

One team member might report they are all green, whilst everyone else reports failures. If someone attempts to fix the issue, they'll go green, while the previous person will not report any failures.

This is really starting to distract us from delivery. Dev's can't honestly say they have completed a task because they cannot add a required acceptance test and get it to pass successfully by the QA team.

Every release requires additional work to get the tests to pass, so at least one member can confirm the acceptance tests are passing.

We are revisiting our acceptance tests to enhance their stability. This includes removing any elements that we deem redundant or duplicated, avoiding the loading of unnecessary data, and utilising programmatic login where possible.

The bulk of that work is being done in [Stabilise the acceptance tests](https://github.com/DEFRA/water-abstraction-acceptance-tests/pull/199). But as the tests rely on this project to tear down and seed test data, there are changes we need to make in support of this.