Updated sending notification to save / fetch and update (#2456)
* Updated sending notification to save / fetch and update

https://eaflood.atlassian.net/browse/WATER-5280

Currently, when a user clicks Confirm on the check-recipients page, as part of sending a notice, we follow this process.

- Convert the notice setup and recipient data into package to be sent to Notify
- We send the package to Notify
- We insert the result and the data into the DB as a ‘notification’.

Because there may be many notifications to send, or in the case of return forms, each can take some time to process, we do this behind the scenes to avoid users seeing any spinning wheels.

So, they will be immediately taken to the ‘Confirmation’ page, which includes a link to view the notice.

If the user is keen, and the process is going to take a moment to complete, when they view the notice and its notifications, the page will be empty. If they were to wait a moment and refresh, then the records would appear. However, we do not consider this a reasonable workaround.

This change updates the notification flow to save a notification, then fetch the notifications and then update the notifications.

We create and fetch the notifications before batching and pass the notifications into the batch process (they will now have a notification id).

The batch process iterates over the notifications and updates (upserts) the saved notification record with Notifies responses.