https://eaflood.atlassian.net/browse/WATER-5280
We are seeing some performance issues before and after updating the notification flow to first save and then fetch the notifications. This was to mark all notification as pending so when the user sees the report page all the notifications are present and marked as pending.

The issue was generating all the PDF for the return forms journey. We here creating the notifications and the PDF before the user could continue to the confirmation page.

When we updated the logic to first save the notification and then fetch, it was simpler to save the PDF as part of that change, knowing we would be generating the PDF just before sending to Notify. So this was a temporary change.

This change implements generating the PDF just before sending to Notify.

This means we no longer generate and save the content before the batch process is started, but we save it as part of the Notify response update.

To achieve this as simply as possible, we have to save the PDF page data into the notification (most of this is already set in the personalisation, but some extra has been added). As this is now in the expected notification format, we need to map the notification data into PDF page data.

This is relevant to mention as we have updated the preview to format the page data through the return forms notifications presenter.