Total amount of water abstracted shown as NaN https://eaflood.atlassian.net/browse/WATER-5291
There is an intermittent bug that is occurring when a return has a start meter reading of zero and a zero meter reading. When submitted, occasionally on the "Abstraction return" page the "Total amount of water abstracted" is being shown as `NaN` instead of zero.

It has also been noticed that the new validation that was added in https://eaflood.atlassian.net/browse/WATER-5192 to prevent all the lines being blank is causing an issue when creating a nil return. This can be fixed by not performing the validation on a nil return as they have no lines.